Define types in a separate file
