+++ src/datasets/geo.dataset.ts
@@ -6,7 +6,7 @@ import { createDefault } from '../utils/preferences';
 import { generateCube } from '../models/data-cube/generation';
 import { Category } from '../models/data-cube/types';
 import { createGeoQuery } from './queries/geo.query';
-import { GeometryCollection, MultiPolygon, Polygon, Topology } from 'topojson-specification';
+import { World } from './geo.types';
 
 export interface Config {
   avgHits: number;
@@ -15,47 +15,6 @@ export interface Config {
   userStdDev: number;
   avgSessionsPerUser: number;
   sessionsPerUserStdDev: number;
-}
-
-export interface City {
-  id: string;
-  countryId: string;
-  subcontinentId: string;
-  continentId: string;
-  name: string;
-  lat: number;
-  lng: number;
-  population: number;
-}
-
-export interface Country {
-  id: string;
-  subcontinentId: string;
-  continentId: string;
-  name: string;
-  cities: Record<string, City>;
-  geometry?: Polygon | MultiPolygon;
-}
-
-export interface Subcontinent {
-  id: string;
-  continentId: string;
-  name: string;
-  countries: Record<string, Country>;
-}
-
-export interface Continent {
-  id: string;
-  name: string;
-  subcontinents: Record<string, Subcontinent>;
-}
-
-export interface World {
-  continents: Record<string, Continent>;
-  subcontinents: Record<string, Subcontinent>;
-  countries: Record<string, Country>;
-  cities: Record<string, City>;
-  topology: Topology<{ land: GeometryCollection }>;
 }
 
 export const configMeta: PreferenceMeta<Config> = {
